requirements.txt
streamlit
python-dotenv
pymupdf
uuid

# LangChain関連
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-ollama
langchain-openai
langchain-text-splitters
langchain-chroma

# Embeddings & LLM
openai
huggingface-hub
sentence-transformers
transformers

# VectorStore
chromadb

# その他
tiktoken
